Restored hps-*-service interfaces...
