+++ README.md
@@ -1,6 +1,6 @@
 ## Heartland Payment Systems Portico SDK for NodeJS
 
-This node.js SDK makes it easy to process payments against the Heartland Payment Systems Gateway.  It integrates the [SecureSubmit](https://developer.heartlandpaymentsystems.com/SecureSubmit/Documentation) functionality as well as providing direct access to the Portico gateway functionality.
+This node.js SDK makes it easy to process payments against the Heartland Payment Systems Gateway.  It integrates the [SecureSubmit](https://developer.heartlandpaymentsystems.com/SecureSubmit/Documentation) functionality as well as providing direct access to the Portico gateway.
 
 ## Installation
 
@@ -10,30 +10,31 @@ This node.js SDK makes it easy to process payments against the Heartland Payment
 
 ```javascript
 
-var portico = require('portico-sdk'),
-	config = {
-		        secretApiKey: 	'skapi_cert_MTyMAQBiHVEAewvIzXVFcmUd2UcyBge_eCpaASUp0A',
-		        publicApiKey: 	'pkapi_cert_jKc1FtuyAydZhZfbB3',
-		        versionNumber: 	'1234',
-		        developerId: 	'123456',
-		        siteTrace: 		'trace0001'
-		    },
+var heartland = require('portico-sdk'),
+    config = {
+                secretApiKey:   'skapi_cert_MTyMAQBiHVEAewvIzXVFcmUd2UcyBge_eCpaASUp0A',
+                publicApiKey:   'pkapi_cert_jKc1FtuyAydZhZfbB3',
+                versionNumber:  '1234',
+                developerId:    '123456',
+                siteTrace:      'trace0001'
+            },
     uri = 'https://cert.api2.heartlandportico.com/Hps.Exchange.PosGateway/PosGatewayService.asmx',
-    secureSubmit = new portico.SecureSubmit(config, uri),
-    porticoReport = new portico.PorticoReport(config, uri);
+    hpsCreditService = new heartland.HpsCreditService(config, uri),
+    porticoReport = new heartland.PorticoReport(config, uri);
 
 // Charging a credit card via the SecureSubmit API...
-secureSubmit.chargeWithCard(amount, currency, card, cardHolder, requestMultiUseToken, memo, callback);
+hpsCreditService.chargeWithCard(amount, currency, card, cardHolder, requestMultiUseToken, memo, callback);
 
 // Access the ReportTxnDetail endpoing directly...
 porticoReport.reportTxnDetail(transactionId, callback);
 
 ```
 
-
-## Implemented Package Functions
+## Implemented Functionality
 
 ### SecureSubmit
+
+#### HpsCreditService
 
 * chargeWithCard
 * chargeWithToken
@@ -46,11 +47,17 @@ porticoReport.reportTxnDetail(transactionId, callback);
 * refundWithTransactionId
 * reverseWithCard
 * reverseWithTransactionId
-* closeBatch
 * get
 * list
 
+#### HpsBatchService
+
+* closeBatch
+
+
 ### PorticoReport
+
+#### ReportService
 
 * reportBatchDetail
 * reportBatchHistory
