+++ index.js
@@ -1,6 +1,9 @@
 'use strict';
 
 module.exports = {
-	SecureSubmit: require('./lib/services/secure-submit'),
-	PorticoReport: require('./lib/services/portico-report')
+    // SecureSubmit Services...
+    HpsCreditService: require('./lib/services/secure-submit/hps-credit-service'),
+    HpsBatchService: require('./lib/services/secure-submit/hps-batch-service'),
+    // Portico Services...
+    PorticoReport: require('./lib/services/portico-services/report-service')
 };
