+++ lib/infrastructure/exception-mapper.js
