+++ lib/services/portico-gateway.js
@@ -2,7 +2,7 @@
 
 var path            = require('path'),
     soap            = require('../soap'),
-    exceptionMapper = require('./exception-mapper');
+    exceptionMapper = require('../infrastructure/exception-mapper');
 
 function PorticoGateway(hpsConfig, soapUri) {
