+++ test/secure-submit-batch-test.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var fs              = require('fs'),
-    config          = require('nconf'),
-    assert          = require('assert'),
-    SecureSubmit    = require('../lib/services/secure-submit');
+var fs               = require('fs'),
+    config           = require('nconf'),
+    assert           = require('assert'),
+    HpsBatchService  = require('../lib/services/secure-submit/hps-batch-service');
 
 if (fs.statSync('./test/config.json')) {
     config.file({file: './test/config.json'});
@@ -11,11 +11,11 @@ if (fs.statSync('./test/config.json')) {
 
 exports.batch_valid_config = {
     setUp: function (callback) {
-        this.secureSubmit   = new SecureSubmit(config.get('validServicesConfig'), config.get('testUri'));
+        this.hpsBatchService = new HpsBatchService(config.get('validServicesConfig'), config.get('testUri'));
         callback();
     },
     closeBatch: function (done) {
-        this.secureSubmit.closeBatch(function (err, result) {
+        this.hpsBatchService.closeBatch(function (err, result) {
             assert.notEqual(result, undefined, 'The result should be something.');
             assert.equal(err, null, 'Should not return an error.');
             done();
