+++ test/secure-submit-credit-test.js
@@ -1,11 +1,11 @@
 'use strict';
 
-var fs             = require('fs'),
-    config         = require('nconf'),
-    assert         = require('assert'),
-    util           = require('util'),
-    https          = require('https'),
-    SecureSubmit   = require('../lib/services/secure-submit');
+var fs               = require('fs'),
+    config           = require('nconf'),
+    assert           = require('assert'),
+    util             = require('util'),
+    https            = require('https'),
+    HpsCreditService = require('../lib/services/secure-submit/hps-credit-service');
 
 if (fs.statSync('./test/config.json')) {
     config.file({file: './test/config.json'});
@@ -36,13 +36,13 @@ function getToken(card, callback) {
 
 exports.credit_valid_config = {
     setUp: function (callback) {
-        this.secureSubmit   = new SecureSubmit(config.get('validServicesConfig'), config.get('testUri'));
+        this.hpsCreditService   = new HpsCreditService(config.get('validServicesConfig'), config.get('testUri'));
         callback();
     },
     list_between_today_and_yesterday: function (done) {
         var startDate = new Date(), endDate = new Date();
         startDate.setDate(startDate.getDate() - 1);
-        this.secureSubmit.list(startDate.toISOString(), endDate.toISOString(), null, function (err, result) {
+        this.hpsCreditService.list(startDate.toISOString(), endDate.toISOString(), null, function (err, result) {
             if (err) return done(err);
             assert.notEqual(result.length, 0, 'The result should be an array with length > 0.');
             assert.equal(err, null, 'Should not return an error.');
@@ -50,7 +50,7 @@ exports.credit_valid_config = {
         });
     },
     get_with_bad_id: function (done) {
-        this.secureSubmit.get(12345, function (err, result) {
+        this.hpsCreditService.get(12345, function (err, result) {
             assert.equal(result, null, 'The result should be null.');
             assert.equal(err.message, 'Report criteria did not produce any results.', 'Should get the correct error message.');
             done();
@@ -59,8 +59,8 @@ exports.credit_valid_config = {
     get_with_good_id: function (done) {
         var startDate = new Date(), endDate = new Date(), that = this;
         startDate.setDate(startDate.getDate() - 1);
-        this.secureSubmit.list(startDate.toISOString(), endDate.toISOString(), null, function (err, result) {
-            that.secureSubmit.get(Number(result[0].transactionId), function (err, result) {
+        this.hpsCreditService.list(startDate.toISOString(), endDate.toISOString(), null, function (err, result) {
+            that.hpsCreditService.get(Number(result[0].transactionId), function (err, result) {
                 if (err) return done(err);
                 assert.notEqual(result, null, 'The result should not be null.');
                 done();
@@ -68,7 +68,7 @@ exports.credit_valid_config = {
         });
     },
     chargeWithValidVisa: function (done) {
-        this.secureSubmit.chargeWithCard(10.00, 'usd', config.get('validVisa'),
+        this.hpsCreditService.chargeWithCard(10.00, 'usd', config.get('validVisa'),
             config.get('validCardHolder'), false, null, function (err, result) {
                 if (err) return done(err);
                 else assert.equal(result.responseCode, '00', 'The response code should be "00".');
@@ -76,7 +76,7 @@ exports.credit_valid_config = {
             });
     },
     chargeWithValidMasterCard: function (done) {
-        this.secureSubmit.chargeWithCard(10.00, 'usd', config.get('validMasterCard'),
+        this.hpsCreditService.chargeWithCard(10.00, 'usd', config.get('validMasterCard'),
             config.get('validCardHolder'), false, null, function (err, result) {
                 if (err) return done(err);
                 else assert.equal(result.responseCode, '00', 'The response code should be "00".');
@@ -86,7 +86,7 @@ exports.credit_valid_config = {
     chargeWithValidMasterCardToken: function (done) {
         var that = this;
         getToken(config.get('validMasterCard'), function (token) {
-            that.secureSubmit.chargeWithToken(10.00, 'usd', token.token_value,
+            that.hpsCreditService.chargeWithToken(10.00, 'usd', token.token_value,
                 config.get('validCardHolder'), false, null, function (err, result) {
                     if (err) return done(err);
                     else assert.equal(result.responseCode, '00', 'The response code should be "00".');
@@ -95,14 +95,14 @@ exports.credit_valid_config = {
         });
     },
     authorizeWithValidMasterCard: function (done) {
-        this.secureSubmit.authorizeWithCard(10.00, 'usd', config.get('validMasterCard'),
+        this.hpsCreditService.authorizeWithCard(10.00, 'usd', config.get('validMasterCard'),
             config.get('validCardHolder'), false, null, function (err, result) {
                 assert.equal(result.responseCode, '00', 'The response code should be "00".');
                 done();
             });
     },
     verifyWithValidMasterCard: function (done) {
-        this.secureSubmit.verifyWithCard(config.get('validMasterCard'),
+        this.hpsCreditService.verifyWithCard(config.get('validMasterCard'),
             config.get('validCardHolder'), false, function (err, result) {
                 assert.equal(result.responseCode, '85', 'The response code should be "85".');
                 done();
@@ -111,7 +111,7 @@ exports.credit_valid_config = {
     verifyWithValidMasterCardToken: function (done) {
         var that = this;
         getToken(config.get('validMasterCard'), function (data) {
-          that.secureSubmit.verifyWithToken(data.token_value,
+          that.hpsCreditService.verifyWithToken(data.token_value,
               config.get('validCardHolder'), false, function (err, result) {
                   assert.equal(result.responseCode, '85', 'The response code should be "85".');
                   done();
@@ -121,7 +121,7 @@ exports.credit_valid_config = {
     verifyWithValidMasterCardTokenGetMulti: function (done) {
         var that = this;
         getToken(config.get('validMasterCard'), function (data) {
-            that.secureSubmit.verifyWithToken(data.token_value,
+            that.hpsCreditService.verifyWithToken(data.token_value,
                 config.get('validCardHolder'), true, function (err, result) {
                     assert.equal(result.responseCode, '85', 'The response code should be "85".');
                     assert.notEqual(result.tokenData, null, 'Multi-use tokenData not available.');
@@ -133,17 +133,17 @@ exports.credit_valid_config = {
     },
     capture: function (done) {
         var that = this;
-        this.secureSubmit.authorizeWithCard(10.00, 'usd', config.get('validMasterCard'),
+        this.hpsCreditService.authorizeWithCard(10.00, 'usd', config.get('validMasterCard'),
             config.get('validCardHolder'), false, null, function (err, result) {
                 assert.equal(result.responseCode, '00', 'The response code should be "00".');
-                that.secureSubmit.capture(result.transactionId, null, function (err, result) {
+                that.hpsCreditService.capture(result.transactionId, null, function (err, result) {
                     assert.equal(result.responseCode, '00', 'The response code should be "00".');
                     done();
                 });
             });
     },
     refundWithValidMasterCard: function (done) {
-        this.secureSubmit.refundWithCard(10.00, 'usd', config.get('validMasterCard'),
+        this.hpsCreditService.refundWithCard(10.00, 'usd', config.get('validMasterCard'),
             config.get('validCardHolder'), null, function (err, result) {
                 assert.notEqual(result.transactionId, undefined, 'The response transaction ID should not be undefined.');
                 done();
@@ -151,10 +151,10 @@ exports.credit_valid_config = {
     },
     refundWithValidTransactionId: function (done) {
         var that = this;
-        this.secureSubmit.chargeWithCard(10.00, 'usd', config.get('validVisa'),
+        this.hpsCreditService.chargeWithCard(10.00, 'usd', config.get('validVisa'),
             config.get('validCardHolder'), false, null, function (err, getResult) {
                 assert.equal(getResult.responseCode, '00', 'The response code should be "00".');
-                that.secureSubmit.refundWithTransactionId(10.00, 'usd', getResult.transactionId,
+                that.hpsCreditService.refundWithTransactionId(10.00, 'usd', getResult.transactionId,
                     config.get('validCardHolder'), null, function (err, refundResult) {
                         assert.notEqual(refundResult.transactionId, undefined, 'The response transaction ID should not be undefined.');
                         done();
@@ -163,10 +163,10 @@ exports.credit_valid_config = {
     },
     reverseWithValidTransactionIdAmountSpecified: function (done) {
         var that = this;
-        this.secureSubmit.chargeWithCard(10.00, 'usd', config.get('validVisa'),
+        this.hpsCreditService.chargeWithCard(10.00, 'usd', config.get('validVisa'),
             config.get('validCardHolder'), false, null, function (err, getResult) {
                 assert.equal(getResult.responseCode, '00', 'The response code should be "00".');
-                that.secureSubmit.reverseWithTransactionId(10.00, 'usd', getResult.transactionId,
+                that.hpsCreditService.reverseWithTransactionId(10.00, 'usd', getResult.transactionId,
                     null, function (err, reverseResult) {
                         assert.notEqual(reverseResult.transactionId, undefined, 'The response transaction ID should not be undefined.');
                         done();
@@ -177,13 +177,13 @@ exports.credit_valid_config = {
 
 exports.credit_invalid_config = {
     setUp: function (callback) {
-        this.secureSubmit   = new SecureSubmit(config.get('invalidServicesConfig'));
+        this.hpsCreditService   = new HpsCreditService(config.get('invalidServicesConfig'));
         callback();
     },
     list_between_today_and_yesterday: function (done) {
         var startDate = new Date(), endDate = new Date();
         startDate.setDate(startDate.getDate() - 1);
-        this.secureSubmit.list(startDate.toISOString(), endDate.toISOString(), null, function (err, result) {
+        this.hpsCreditService.list(startDate.toISOString(), endDate.toISOString(), null, function (err, result) {
             assert.notEqual(err.message, null, 'An error should be thrown indicating an authentication problem.');
             done();
         });
